agentic_ai/pyproject.toml
[project]
name = "agentic_ai"
version = "0.1.0"
description = "Agentic AI using CrewAI and Ollama"
authors = [
    { name = "Your Name", email = "you@example.com" }
]
readme = "README.md"
requires-python = ">=3.10,<3.14"
dependencies = [
    "crewai[tools]>=0.177.0,<1.0.0",
    "crewai-tools",
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    "matplotlib>=3.7.0",
    "seaborn>=0.12.0",
    "ollama>=0.1.8",
    "python-dotenv>=1.0.0",
    "tabulate>=0.9.0",
    "rich>=13.0.0",
    "pyyaml>=6.0.0",
]

[project.scripts]
agentic_ai = "agentic_ai.main:run"
run_crew = "agentic_ai.main:run"
train = "agentic_ai.main:train"
replay = "agentic_ai.main:replay"
test = "agentic_ai.main:test"

[tool.crewai]
type = "crew"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

